+++ CHANGELOG.md
@@ -5,8 +5,8 @@
 * by default, cookies are not encrypted;
 * `AppTable::getList()`, `AppTable::getTreeList()` and
     `UsersTable::getActiveList()` now return a `Query` object;
-* added tests for `BannersController` and `BannersPositionsController` admin
-    classes.
+* added tests for `BannersController`, `BannersPositionsController` and
+    `LogsController` admin classes.
 
 
 ## 2.17 branch